Organize hooks into subfolders
