+++ src/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from "./useAccountProfile"
 export * from "./useDisclosure"
 export * from "./useTitle"
-export * from "./useSitemap"
 export * from "./useBreadcrumbs"
 export * from "./useLinkShims"
 export * from "./useFormHelpers"
