+++ src/hooks/sitemap/index.ts
@@ -1,0 +1,2 @@
+export * from "./useSitemap"
+export * from "./useSitemapGroups"
